+++ examples/menu/discords/AppDiscord.ts
@@ -4,7 +4,11 @@ import {
   MessageSelectMenu,
   SelectMenuInteraction,
 } from "discord.js";
-import { Discord, SelectMenuComponent, Slash } from "../../../build/cjs/index.cjs";
+import {
+  Discord,
+  SelectMenuComponent,
+  Slash,
+} from "../../../build/cjs/index.cjs";
 
 const roles = [
   { label: "Principal", value: "principal" },
