+++ packages/music/examples/LavalinkPlayer/discords/AppDiscord.ts
@@ -1,7 +1,7 @@
 import { Client, Discord, On, Slash, SlashOption } from "discordx";
 import { CommandInteraction, GuildMember } from "discord.js";
 import type { ArgsOf } from "discordx";
-import { Lava } from "../../../src/index.js";
+import { Lava } from "../../../build/cjs/index.cjs";
 
 @Discord()
 export class music {
