+++ packages/utilities/examples/pagination/Main.ts
@@ -1,7 +1,7 @@
 import "reflect-metadata";
-import { dirname, importx } from "@discordx/importer";
 import { Client } from "discordx";
 import { Intents } from "discord.js";
+import { importx } from "@discordx/importer";
 
 export class Main {
   private static _client: Client;
@@ -33,7 +33,7 @@ export class Main {
       this._client.executeInteraction(interaction);
     });
 
-    await importx(dirname(import.meta.url) + "/discords/**/*.{js,ts}");
+    await importx(__dirname + "/discords/**/*.{js,ts}");
     await this._client.login(process.env.BOT_TOKEN ?? "");
   }
 }
