+++ packages/utilities/examples/pagination/discords/slash command info.ts
@@ -1,6 +1,6 @@
 import { CommandInteraction, MessageEmbed } from "discord.js";
 import { Discord, MetadataStorage, Slash } from "discordx";
-import { Pagination } from "../../../src/index.js";
+import { Pagination } from "../../../build/cjs/index.cjs";
 
 @Discord()
 export abstract class SlashExample {
